Create one folder per sample
Created the possibility to have one folder per settings (file), such that one soesn't need to change the code just for running another case.
This needed some changes in the code in order to find the executiong directory.

I had lot's of trouble to get it working together with the last lines for the Analyze.C script. For some reason the program almost refused to collaborate, and just arbitrarly either ran, or gave non-sense output. Now fixed.