Fixes #5 -- Parity dependent Jpop
Add new parameter, bParPop_Equipar, to determine, whether RAINIER shall use i) equiparity for the Jpop file or ii) read a parity dependent file.

In the first case, which was the default up to now, the population file with the output from TALYS is the same as before. For the second version, one has to give a file with a header of the type:
" bin    Ex    Popul.    J= 0.0-   J= 0.0+   J= 1.0-   J= 1.0+ ..."

This can be fixed eg. in the in the binary.f and/or multiple.f rutine of talys.