Solves Bug #12 reading discrete levels
Slight changes in how the discrete states are read. Solves bugs when reading i) unstable isotopes or ii) if halflive is very large.
Now it throws a warning if the halflive of a state is not given (and sets it to a quasi stable state. However, the best might be to "simply" change the levels file in this cases with reasonable values.

In connection with this, I implemented the possibility to read a modified level file. It should be in the same folder; see naming convention in the code.