Bugfix for Wigner distribution
Wigner distribution was set up incorrect. If it was chosen, an erroneous
level scheme was created, most visible when the expected number of
levels at Ecrit (so the first continuum bin for each Jpi) was low.

Error: Number of levels in each 1st Jpi bin was set to 1 at minimum.

Fix: rearranged the initialization of `Wignersum` to match Dicebox.
(Honestly I still haven't understood the random number drawing there
100% yet, but it's accepted code).